move most p2p code out of main into net_processing
