+++ src/test/DoS_tests.cpp
@@ -10,6 +10,7 @@
 #include "keystore.h"
 #include "main.h"
 #include "net.h"
+#include "net_processing.h"
 #include "pow.h"
 #include "script/sign.h"
 #include "serialize.h"
