+++ src/test/exploit_tests.cpp
@@ -8,6 +8,7 @@
 #include "dosman.h"
 #include "main.h"
 #include "net.h"
+#include "net_processing.h"
 #include "primitives/block.h"
 #include "protocol.h"
 #include "random.h"
